feat: add base agent class (#922)
