+++ packages/core/src/storage/migration/UpdateAssistant.ts
@@ -1,8 +1,7 @@
-import type { Agent } from '../../agent/Agent'
+import type { BaseAgent } from '../../agent/BaseAgent'
 import type { FileSystem } from '../FileSystem'
 import type { UpdateConfig } from './updates'
 
-import { AgentContext } from '../../agent'
 import { InjectionSymbols } from '../../constants'
 import { AriesFrameworkError } from '../../error'
 import { isFirstVersionHigherThanSecond, parseVersionString } from '../../utils/version'
@@ -12,11 +11,10 @@ import { StorageUpdateService } from './StorageUpdateService'
 import { StorageUpdateError } from './error/StorageUpdateError'
 import { CURRENT_FRAMEWORK_STORAGE_VERSION, supportedUpdates } from './updates'
 
-export class UpdateAssistant {
+export class UpdateAssistant<Agent extends BaseAgent = BaseAgent> {
   private agent: Agent
   private storageUpdateService: StorageUpdateService
   private updateConfig: UpdateConfig
-  private agentContext: AgentContext
   private fileSystem: FileSystem
 
   public constructor(agent: Agent, updateConfig: UpdateConfig) {
@@ -24,7 +22,6 @@ export class UpdateAssistant {
     this.updateConfig = updateConfig
 
     this.storageUpdateService = this.agent.dependencyManager.resolve(StorageUpdateService)
-    this.agentContext = this.agent.dependencyManager.resolve(AgentContext)
     this.fileSystem = this.agent.dependencyManager.resolve<FileSystem>(InjectionSymbols.FileSystem)
   }
 
@@ -46,11 +43,11 @@ export class UpdateAssistant {
   }
 
   public async isUpToDate() {
-    return this.storageUpdateService.isUpToDate(this.agentContext)
+    return this.storageUpdateService.isUpToDate(this.agent.context)
   }
 
   public async getCurrentAgentStorageVersion() {
-    return this.storageUpdateService.getCurrentStorageVersion(this.agentContext)
+    return this.storageUpdateService.getCurrentStorageVersion(this.agent.context)
   }
 
   public static get frameworkStorageVersion() {
@@ -59,7 +56,7 @@ export class UpdateAssistant {
 
   public async getNeededUpdates() {
     const currentStorageVersion = parseVersionString(
-      await this.storageUpdateService.getCurrentStorageVersion(this.agentContext)
+      await this.storageUpdateService.getCurrentStorageVersion(this.agent.context)
     )
 
     // Filter updates. We don't want older updates we already applied
@@ -113,7 +110,7 @@ export class UpdateAssistant {
           await update.doUpdate(this.agent, this.updateConfig)
 
           // Update the framework version in storage
-          await this.storageUpdateService.setCurrentStorageVersion(this.agentContext, update.toVersion)
+          await this.storageUpdateService.setCurrentStorageVersion(this.agent.context, update.toVersion)
           this.agent.config.logger.info(
             `Successfully updated agent storage from version ${update.fromVersion} to version ${update.toVersion}`
           )
