+++ packages/core/src/storage/migration/updates/0.1-0.2/index.ts
@@ -1,4 +1,4 @@
-import type { Agent } from '../../../../agent/Agent'
+import type { BaseAgent } from '../../../../agent/BaseAgent'
 import type { UpdateConfig } from '../../updates'
 
 import { migrateConnectionRecordToV0_2 } from './connection'
@@ -9,7 +9,7 @@ export interface V0_1ToV0_2UpdateConfig {
   mediationRoleUpdateStrategy: 'allMediator' | 'allRecipient' | 'recipientIfEndpoint' | 'doNotChange'
 }
 
-export async function updateV0_1ToV0_2(agent: Agent, config: UpdateConfig): Promise<void> {
+export async function updateV0_1ToV0_2<Agent extends BaseAgent>(agent: Agent, config: UpdateConfig): Promise<void> {
   await migrateCredentialRecordToV0_2(agent)
   await migrateMediationRecordToV0_2(agent, config.v0_1ToV0_2)
   await migrateConnectionRecordToV0_2(agent)
